+++ engine/engine/MonsterStats.hpp
@@ -7,8 +7,8 @@ enum class MonsterType : uint8_t;
 class MonsterStats
 {
 public:
-  MonsterStats(MonsterType type, int level, uint8_t dungeonMultiplier);
-  MonsterStats(int level, int hpMax, int damage, int deathProtection);
+  MonsterStats(MonsterType type, uint8_t level, uint8_t dungeonMultiplier);
+  MonsterStats(uint8_t level, uint16_t hpMax, uint16_t damage, uint8_t deathProtection);
 
   MonsterType getType() const;
   uint8_t getLevel() const;
@@ -18,8 +18,8 @@ public:
   uint16_t getHitPoints() const;
   uint16_t getHitPointsMax() const;
 
-  void healHitPoints(int amountPointsHealed, bool allowOverheal);
-  void loseHitPoints(int amountPointsLost);
+  void healHitPoints(uint16_t amountPointsHealed, bool allowOverheal);
+  void loseHitPoints(uint16_t amountPointsLost);
   void setHitPointsMax(uint16_t newHitPointsMax);
 
   uint16_t getDamage() const;
